+++ src/ska_tmc_dishleafnode/manager/component_manager.py
@@ -40,7 +40,7 @@ from ska_tmc_common.v1.tmc_component_manager import TmcLeafNodeComponentManager
 
 from ska_tmc_dishleafnode.az_el_converter import AzElConverter
 from ska_tmc_dishleafnode.commands import (
-    AbortCommands,
+    Abort,
     ApplyPointingModel,
     Configure,
     ConfigureBand,
@@ -221,6 +221,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             "Scan",
             "TrackLoadStaticOff",
             "TrackStop",
+            "Abort",
         )
 
         # Event Receiver
@@ -311,6 +312,12 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
                 "status": None,
             }
             self.end_scan_result = {
+                "result_code": None,
+                "message": None,
+                "exception": None,
+                "status": None,
+            }
+            self.abort_result = {
                 "result_code": None,
                 "message": None,
                 "exception": None,
@@ -699,8 +706,8 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             timestamp = self.convert_timestamp(timestamp_tai_ska_epoch)
             if timestamp:
                 right_ascension, declination = self.converter.azel_to_radec(
-                    str(azimuth),
-                    str(elevation),
+                    azimuth,
+                    elevation,
                     timestamp,
                 )
                 self.actual_pointing = [
@@ -1171,23 +1178,98 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         )
         return task_status, response
 
-    # pylint: disable=arguments-differ
-    def abort_commands(
-        self, task_callback: TaskCallbackType
-    ) -> Tuple[TaskStatus, str]:
+    def abort(self, task_callback: TaskCallbackType) -> Tuple:
         """
-        Invokes Abort command on Dish manager.
+        Abort the Dish.
+
+        :return: a result code and message
         """
-        abort_command = AbortCommands(
-            self,
-            adapter_factory=self.adapter_factory,
-            logger=self.logger,
-        )
-        task_status, response = self.submit_task(
-            abort_command.invoke_abort,
-            task_callback=task_callback,
-        )
-        return task_status, response
+        # base classes set and clear immediately, so we set to
+        # clear ongoing observers and timers.
+        self.abort_event.set()
+        self.observable.notify_observers(attribute_value_change=True)
+        self.abort_event.clear()
+
+        def _invoke_abort_callback(
+            status=None,
+            progress=None,
+            result=None,
+            exception=None,
+        ):
+            """
+            Method for invoking abort callback
+
+            :param status: Status of the task
+            :type status: TaskStatus
+            :param progress: progress of the task
+            :type progress: int
+            :param result: JSON serializable result of the task
+            :type result: Any
+            :param exception : exception raised from the task
+            :type exception: Exception
+
+            :return: None
+
+            """
+            # progress completed in the base class is assumed to be 50%
+            progress_completed = 50
+            if progress is not None:
+                task_callback(progress=progress_completed + progress / 2)
+
+            if status == TaskStatus.FAILED:
+                task_callback(
+                    status=status, exception=exception, result=result
+                )
+            elif status == TaskStatus.COMPLETED:
+                task_callback(status=status, progress=100, result=result)
+
+        # pylint: disable=unused-argument
+        def _abort_commands_callback(
+            status=None,
+            progress=None,
+            result=None,
+            exception=None,
+        ):
+            """
+            Method for abort command callback
+
+            :param status: Status of the task
+            :type status: TaskStatus
+            :param progress: progress of the task
+            :type progress: int
+            :param result: JSON serializable result of the task
+            :type result: Any
+            :param exception : exception raised from the task
+            :type exception: Exception
+
+            :return: None
+            """
+
+            if progress is not None:
+                task_callback(progress=progress / 2)
+            if status == TaskStatus.FAILED:
+                task_callback(
+                    status=status, exception="Failed to abort commands"
+                )
+            elif status == TaskStatus.COMPLETED:
+                task_callback(
+                    progress=50,
+                )
+                abort_command = Abort(
+                    self,
+                    self.op_state_model,
+                    self.adapter_factory,
+                    logger=self.logger,
+                )
+                # Send dummy task_abort_event
+                abort_command.invoke_abort(
+                    task_callback=_invoke_abort_callback,
+                    task_abort_event=threading.Event(),
+                )
+
+        # pylint: enable=unused-argument
+        self.command_in_progress = "Abort"
+        return self.abort_tasks(task_callback=_abort_commands_callback)
 
     def is_trackloadstaticoff_allowed(self: DishLNComponentManager) -> bool:
         """Checks if the command TrackLoadStaticOff is allowed.
@@ -1563,9 +1645,19 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
                 "PointingState value updated to "
                 + f"{PointingState(pointingState).name}"
             )
+
+            # It was observed that DishMaster sends Track pointingState
+            # after configure command is completed on TMC Subarray which
+            # causes issues in aggregation of Subarray Node.
+            # In order to avoid same below filtering criteria has been applied.
+
+            if self.is_configure_command or self.command_in_progress:
+                if self._update_pointingstate_callback:
+                    self._update_pointingstate_callback(pointingState)
+            else:
+                self.logger.debug("pointingState event not pushed")
+
             self.observable.notify_observers(attribute_value_change=True)
-            if self._update_pointingstate_callback:
-                self._update_pointingstate_callback(pointingState)
 
     def update_device_configured_band(
         self: DishLNComponentManager, configured_band: Band
@@ -1596,7 +1688,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
             0
         ]  # station names in the layout json are in capital
 
-    def is_abortcommands_allowed(self: DishLNComponentManager) -> bool:
+    def is_abort_allowed(self: DishLNComponentManager) -> bool:
         """
         Checks whether this command is allowed
         It checks that the device is in the right state
@@ -1606,7 +1698,7 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         :return: True if this command is allowed
         :rtype: boolean
         """
-        # dish manager allows abortcommands in all the dish modes
+        # dish manager allows abort in all the dish modes
         # and pointing states
         # TO DO: DishMode/s & pointing state/s decision
 
@@ -1870,8 +1962,17 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
                             "Track result: %s",
                             self.track_result,
                         )
+
                         notify = True
                         self.logger.info("Done Track update")
+                    elif "Abort" in unique_id:
+                        self.abort_result["result_code"] = result_code
+                        self.abort_result["message"] = message
+                        self.logger.debug(
+                            "Abort result: %s",
+                            self.abort_result,
+                        )
+                        notify = True
 
             self.logger.info("Released command_result_update_lock")
             if notify:
@@ -1879,6 +1980,8 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
                 self.observable.notify_observers(attribute_value_change=True)
             else:
                 self.logger.info("Notify not set")
+
+                self.observable.notify_observers(attribute_value_change=True)
 
             if result_code in [
                 ResultCode.FAILED,
@@ -2255,6 +2358,26 @@ class DishLNComponentManager(TmcLeafNodeComponentManager):
         """
         with self.command_result_update_lock:
             return self.set_operate_mode_result
+
+    def get_abort_result_code(self: DishLNComponentManager) -> ResultCode:
+        """
+        Return the result of the Abort command execution
+
+        :return: ResultCode from the set_abort_result
+        :rtype: ResultCode
+        """
+        with self.command_result_update_lock:
+            return self.abort_result["result_code"]
+
+    def get_abort_result_dict(self: DishLNComponentManager) -> dict:
+        """
+        Return the dictinary containing Abort command execution status
+
+        :return: abort_result dictionary
+        :rtype: dict
+        """
+        with self.command_result_update_lock:
+            return self.abort_result
 
     def get_track_result_code(self: DishLNComponentManager):
         """