+++ tests/integration/test_dishln_scan_command.py
@@ -51,10 +51,6 @@ def scan_command_timeout(
         lookahead=2,
     )
 
-    group_callback["pointingState"].assert_change_event(
-        (PointingState.NONE),
-        lookahead=4,
-    )
     result_fp, unique_id_fp = dish_leaf_node.SetStandbyFPMode()
     assert result_fp[0] == ResultCode.QUEUED
 
@@ -185,10 +181,6 @@ def scan_command_error_propagation(
         lookahead=2,
     )
 
-    group_callback["pointingState"].assert_change_event(
-        (PointingState.NONE),
-        lookahead=4,
-    )
     result_fp, unique_id_fp = dish_leaf_node.SetStandbyFPMode()
     assert result_fp[0] == ResultCode.QUEUED
 
@@ -318,10 +310,6 @@ def scan_command(
         lookahead=2,
     )
 
-    group_callback["pointingState"].assert_change_event(
-        (PointingState.NONE),
-        lookahead=4,
-    )
     result_fp, unique_id_fp = dish_leaf_node.SetStandbyFPMode()
     assert result_fp[0] == ResultCode.QUEUED