+++ tests/unit/test_device_unresponsive.py
@@ -12,7 +12,7 @@ def test_abort_command_fail_check_allowed_with_device_unresponsive(
     with pytest.raises(
         DeviceUnresponsive, match=f"{DISH_MASTER_DEVICE} not available"
     ):
-        cm.is_abortcommands_allowed()
+        cm.is_abort_allowed()
 
 
 def test_trackloadstaticoff_command_not_allowed(cm_without_er_lp):